requirements.txt
langchain_experimental==0.0.57
langchain_openai==0.1.6
lida==0.0.14
matplotlib==3.8.4
openai==1.25.1
pandas==1.5.3
Pillow==10.3.0
plotly==5.21.0
python-dotenv==1.0.1
streamlit==1.33.0
tabulate==0.9.0
